+++ src/preferences/settings-window/tabs/UpgradeTab.swift
@@ -1,3 +1,7 @@
+// Starting 2026-06-11, Ky forked the original repo to make this one.
+// The details of changes to this file (and all other files in this repository), including when the changes were made, can be found in the Git metadata of this repository.
+// If you receive a version of this repository that is lacking the Git metadata, you may contact Ky and they will provide that metadata to you free of charge: FreeAltTab@KyNorthstar.me
+
 import Cocoa
 
 class UpgradeTab {
@@ -113,7 +117,7 @@ class UpgradeTab {
 
     private static func makeHeaderStrip() -> NSStackView {
         let titleFont = NSFont.systemFont(ofSize: 15, weight: .medium)
-        let titleText = NSLocalizedString("AltTab Pro", comment: "")
+        let titleText = NSLocalizedString("FreeAltTab Pro", comment: "")
         let titleAttr = NSMutableAttributedString(string: titleText, attributes: [
             .font: titleFont,
             .foregroundColor: NSColor.labelColor,
@@ -172,7 +176,7 @@ class UpgradeTab {
         table.addRow(leftText: NSLocalizedString("Manage activations, view receipts, etc", comment: ""), rightViews: [manageButton], isAddSeparator: false)
         let deactivateButton = NSButton(title: NSLocalizedString("Deactivate license", comment: ""), target: nil, action: nil)
         deactivateButton.onAction = { _ in deactivateLicense() }
-        let deactivateHint = TableGroupView.makeText(NSLocalizedString("License will remain valid and usable to activate AltTab on any machine", comment: ""))
+        let deactivateHint = TableGroupView.makeText(NSLocalizedString("License will remain valid and usable to activate FreeAltTab on any machine", comment: ""))
         deactivateHint.textColor = .secondaryLabelColor
         table.addRow(leftViews: [TableGroupView.makeText(NSLocalizedString("Deactivate license on this machine", comment: ""))], rightViews: [deactivateButton], secondaryViews: [deactivateHint])
         return table