+++ src/pro/scheduling/Day21ReminderPopover.swift
@@ -1,3 +1,7 @@
+// Starting 2026-06-11, Ky forked the original repo to make this one.
+// The details of changes to this file (and all other files in this repository), including when the changes were made, can be found in the Git metadata of this repository.
+// If you receive a version of this repository that is lacking the Git metadata, you may contact Ky and they will provide that metadata to you free of charge: FreeAltTab@KyNorthstar.me
+
 import Cocoa
 
 class Day21ReminderPopover {
@@ -11,7 +15,7 @@ class Day21ReminderPopover {
         let container = NSView()
         container.translatesAutoresizingMaskIntoConstraints = false
 
-        let title = NSTextField(labelWithString: NSLocalizedString("AltTab Pro is still available", comment: ""))
+        let title = NSTextField(labelWithString: NSLocalizedString("FreeAltTab Pro is still available", comment: ""))
         title.font = .systemFont(ofSize: 13, weight: .semibold)
         title.translatesAutoresizingMaskIntoConstraints = false