+++ src/pro/ui/ProPromptWindow.swift
@@ -1,3 +1,7 @@
+// Starting 2026-06-11, Ky forked the original repo to make this one.
+// The details of changes to this file (and all other files in this repository), including when the changes were made, can be found in the Git metadata of this repository.
+// If you receive a version of this repository that is lacking the Git metadata, you may contact Ky and they will provide that metadata to you free of charge: FreeAltTab@KyNorthstar.me
+
 import Cocoa
 
 /// Base class for non-modal Pro-transition windows ([A] Welcome, [C] Full Upgrade, [D] Proactive,
@@ -9,7 +13,7 @@ class ProPromptWindow: NSWindow {
         var mask: NSWindow.StyleMask = [.titled, .closable, .fullSizeContentView]
         if miniaturizable { mask.insert(.miniaturizable) }
         self.init(contentRect: NSRect(origin: .zero, size: size), styleMask: mask, backing: .buffered, defer: false)
-        title = NSLocalizedString("AltTab Pro", comment: "")
+        title = NSLocalizedString("FreeAltTab Pro", comment: "")
         titleVisibility = .hidden
         titlebarAppearsTransparent = true
         standardWindowButton(.closeButton)?.isHidden = true